+++ algo/src/resource.rs
@@ -1,38 +1,28 @@
 use std::{
   hash::{Hash, Hasher},
   ops::Deref,
-  rc::Rc,
 };
 
 use serde::{Deserialize, Serialize};
-
-/// Enum to store both share and local resource. Share resources can have many
-/// copy, and every copy as the same one. Local resources means is individual
-/// and no other copy exist, two local resources will not as same one even its
-/// content is equal.
-#[derive(Eq, Debug, Clone)]
-pub enum Resource<T> {
-  Share(ShareResource<T>),
-  Local(T),
-}
 
 /// A smarter pointer help us to share resource in application, and it' use a
 /// cheap way to compare if two resource is same one.
 ///
 /// # Notice
-/// Compare two `ShareResource` just compare if it come form same resource not
+/// Compare two `Resource` just compare if it come form same resource not
 /// compare its content.
 #[derive(Debug, Serialize, Deserialize)]
-pub struct ShareResource<T>(Rc<T>);
+pub struct Resource<T>(triomphe::Arc<T>);
 
-impl<T> ShareResource<T> {
+impl<T> Resource<T> {
   #[inline]
-  pub fn new(v: T) -> Self { ShareResource(Rc::new(v)) }
+  pub fn new(v: T) -> Self { Resource(triomphe::Arc::new(v)) }
+
   #[inline]
-  pub fn as_ptr(this: &Self) -> *const () { Rc::as_ptr(&this.0) as *const () }
+  pub fn as_ptr(this: &Self) -> *const () { triomphe::Arc::as_ptr(&this.0) as *const () }
 }
 
-impl<T> Clone for ShareResource<T> {
+impl<T> Clone for Resource<T> {
   #[inline]
   fn clone(&self) -> Self { Self(self.0.clone()) }
 }
@@ -40,50 +30,19 @@ impl<T> Clone for ShareResource<T> {
 impl<T> Deref for Resource<T> {
   type Target = T;
 
-  fn deref(&self) -> &Self::Target {
-    match self {
-      Resource::Share(s) => s.deref(),
-      Resource::Local(l) => l,
-    }
-  }
+  fn deref(&self) -> &Self::Target { self.0.deref() }
 }
 
 impl<T> PartialEq for Resource<T> {
   #[inline]
-  fn eq(&self, other: &Self) -> bool {
-    match (self, other) {
-      (Resource::Share(a), Resource::Share(b)) => a == b,
-      _ => false,
-    }
-  }
+  fn eq(&self, other: &Self) -> bool { triomphe::Arc::ptr_eq(&self.0, &other.0) }
 }
 
-impl<T> Deref for ShareResource<T> {
-  type Target = T;
+impl<T> Eq for Resource<T> {}
 
-  fn deref(&self) -> &Self::Target { self.0.deref() }
-}
-
-impl<T> PartialEq for ShareResource<T> {
+impl<T> Hash for Resource<T> {
   #[inline]
-  fn eq(&self, other: &Self) -> bool { Rc::ptr_eq(&self.0, &other.0) }
-}
-
-impl<T> Eq for ShareResource<T> {}
-
-impl<T> Hash for ShareResource<T> {
-  #[inline]
-  fn hash<H: Hasher>(&self, state: &mut H) { Rc::as_ptr(&self.0).hash(state); }
-}
-
-impl<T> From<T> for Resource<T> {
-  #[inline]
-  fn from(r: T) -> Self { Resource::Local(r) }
-}
-
-impl<T> From<ShareResource<T>> for Resource<T> {
-  #[inline]
-  fn from(s: ShareResource<T>) -> Self { Resource::Share(s) }
+  fn hash<H: Hasher>(&self, state: &mut H) { triomphe::Arc::as_ptr(&self.0).hash(state); }
 }
 
 #[cfg(test)]
@@ -92,8 +51,8 @@ mod tests {
 
   #[test]
   fn compare() {
-    let a = ShareResource::new(5);
-    let b = ShareResource::new(5);
+    let a = Resource::new(5);
+    let b = Resource::new(5);
     assert_ne!(a, b);
 
     #[allow(clippy::redundant_clone)]
@@ -104,19 +63,18 @@ mod tests {
 
   #[test]
   fn hash() {
-    let a = ShareResource::new("a");
+    let a = Resource::new("a");
     let mut map = std::collections::HashSet::new();
     map.insert(a.clone());
 
-    assert!(!map.contains(&ShareResource::new("a")));
+    assert!(!map.contains(&Resource::new("a")));
     assert!(map.contains(&a));
   }
 
   #[test]
   fn share_local_compare() {
-    let s = ShareResource::new(1);
-    let l = Resource::Local(1);
-    assert_ne!(l, s.into());
-    assert_ne!(l, 1.into());
+    let a = Resource::new(1);
+    let b = Resource::new(1);
+    assert_ne!(b, a);
   }
 }